Add timing to the log statements
Normally you could also get timings form aiohttp.server's built-in
response logging, but that doesn't seem to be enabled in the deployment.
This commit as a simple Timer context manager and wraps the responses
in that context and adds the timing as a duration property on the log
statement. This will make it easier to monitor the request rate and how
long the responses are taking to compute in operations.